refactor: api search routes
